BUG#: 1757 TITLE: Fixed warning about comparison of unsigned value < 0 always true.
DESCRIPTION: There was an issue with how _itoa() works.  Corrected with a cast.

+++ src/slp/slp_client/src/cmd-utils/slp_client/lslp-linux.h
@@ -26,7 +26,7 @@
  *	Original Author: Mike Day md@soft-hackle.net
  *                                mdd@us.ibm.com
  *
- *  $Header: /cvs/MSB/pegasus/src/slp/slp_client/src/cmd-utils/slp_client/lslp-linux.h,v 1.2 2004/06/03 23:22:30 tony Exp $ 	                                                            
+ *  $Header: /cvs/MSB/pegasus/src/slp/slp_client/src/cmd-utils/slp_client/lslp-linux.h,v 1.3 2004/07/30 16:48:44 david.dillard Exp $ 	                                                            
  *               					                    
  *  Copyright (c) 2001 - 2003  IBM                                          
  *  Copyright (c) 2000 - 2003 Michael Day                                    
@@ -193,10 +193,10 @@ void  hug_num_to_ascii(uint64 val, int8 *buf, int32 radix, BOOL is_neg);
 #endif
 
 #define _ultoa(v, b, r) num_to_ascii((uint32)(v), (b), (r), FALSE)
-#define _itoa(v, b, r) num_to_ascii((uint32)(v), (b), (r), (((r) == 10) && ((v) < 0)))
-#define _ltoa(v, b, r) num_to_ascii((uint32)(v), (b), (r), (((r) == 10) && ((v) < 0)))
+#define _itoa(v, b, r) num_to_ascii((uint32)(v), (b), (r), (((r) == 10) && ((int32)(v) < 0)))
+#define _ltoa(v, b, r) num_to_ascii((uint32)(v), (b), (r), (((r) == 10) && ((int32)(v) < 0)))
 #define _ul64toa(v, b, r) huge_num_to_ascii((uint64)(v), (b), (r), FALSE)
-#define _i64toa(v, b, r) huge_num_to_ascii((uint64)(v), (b), (r), (((r) == 10) && ((v) < 0)))
+#define _i64toa(v, b, r) huge_num_to_ascii((uint64)(v), (b), (r), (((r) == 10) && ((int64)(v) < 0)))
